+++ smart-rent/src/main/java/com/smartrent/infra/repository/entity/Listing.java
@@ -185,6 +185,9 @@ public class Listing {
     @CreationTimestamp
     LocalDateTime createdAt;
 
+    @Column(name = "pushed_at")
+    LocalDateTime pushedAt;
+
     @Column(name = "updated_at")
     @UpdateTimestamp
     LocalDateTime updatedAt;
