+++ smart-rent/src/main/java/com/smartrent/service/push/PushService.java
@@ -1,8 +1,5 @@
 package com.smartrent.service.push;
 
-import com.smartrent.dto.request.BoostListingRequest;
-import com.smartrent.dto.request.ScheduleBoostRequest;
-import com.smartrent.dto.response.BoostResponse;
 import com.smartrent.infra.repository.entity.PushHistory;
 import com.smartrent.infra.repository.entity.PushSchedule;
 
@@ -11,74 +8,13 @@ import java.time.LocalTime;
 import java.util.List;
 
 /**
- * Service interface for managing listing push/boost operations.
- * Handles pushing listings (boost) immediately, scheduling automatic pushes, and recording push history.
+ * Service interface for managing listing push operations.
+ * Handles pushing listings at scheduled times and recording push history.
  */
 public interface PushService {
 
     /**
-     * Boost (push) a listing immediately.
-     * Can use membership quota or require payment.
-     *
-     * @param userId The user ID
-     * @param request The boost request containing listing ID and payment preferences
-     * @return BoostResponse with result or payment URL if payment required
-     */
-    BoostResponse boostListing(String userId, BoostListingRequest request);
-
-    /**
-     * Complete boost after successful payment.
-     * Called from payment callback handler.
-     *
-     * @param transactionId The completed transaction ID
-     * @return BoostResponse with result
-     */
-    BoostResponse completeBoostAfterPayment(String transactionId);
-
-    /**
-     * Schedule automatic boosts for a listing.
-     * Can use membership quota or require payment.
-     *
-     * @param userId The user ID
-     * @param request The schedule request containing listing ID, time, and total pushes
-     * @return BoostResponse with schedule result
-     */
-    BoostResponse scheduleBoost(String userId, ScheduleBoostRequest request);
-
-    /**
-     * Execute all scheduled boosts for the current time.
-     * Called by cron job every minute/hour.
-     *
-     * @return Number of boosts executed
-     */
-    int executeScheduledBoosts();
-
-    /**
-     * Cancel a scheduled boost.
-     *
-     * @param userId The user ID
-     * @param scheduleId The schedule ID to cancel
-     */
-    void cancelScheduledBoost(String userId, Long scheduleId);
-
-    /**
-     * Get boost/push history for a listing.
-     *
-     * @param listingId The listing ID
-     * @return List of boost responses
-     */
-    List<BoostResponse> getBoostHistory(Long listingId);
-
-    /**
-     * Get boost/push history for all listings owned by a user.
-     *
-     * @param userId The user ID
-     * @return List of boost responses for all user's listings
-     */
-    List<BoostResponse> getUserBoostHistory(String userId);
-
-    /**
-     * Create a push schedule for a listing.
+     * Create a new push schedule for a listing.
      * Validates that listing doesn't already have an active schedule.
      *
      * @param listingId The ID of the listing to schedule
