+++ smart-rent/src/main/resources/application-local.yaml
@@ -1,7 +1,7 @@
 #DB
 IDENTITY_SERVICE_DB_URL: "jdbc:mysql://localhost:3306/smartrent"
 DB_USERNAME: root
-DB_PASSWORD: vuhuydiet
+DB_PASSWORD: root
 
 #Redis
 REDIS_USERNAME: ""
